+++ src/altimeter/led.h
@@ -1,7 +1,7 @@
 #ifndef __in_led_h
 #define __in_led_h
 
-void LED_show(byte red, byte green, byte blue, int delayMs = 0);
+void LED_show(byte red, byte green, byte blue, uint8_t delayMs = 0);
 
 bool IsPWMActive();
